Clean up API definitions
