+++ skyplane/__init__.py
@@ -1,10 +1,9 @@
 from pathlib import Path
 
 from skyplane import compute
-from skyplane.api.auth_config import AWSConfig, AzureConfig, GCPConfig
 from skyplane.api.client import SkyplaneClient
-from skyplane.api.dataplane import Dataplane
-from skyplane.api.transfer_config import TransferConfig
+from skyplane.api.config import TransferConfig, AWSConfig, AzureConfig, GCPConfig
+from skyplane.api.provision.dataplane import Dataplane
 
 __version__ = "0.2.1"
 __root__ = Path(__file__).parent.parent
