+++ skyplane/api/provision/dataplane.py
@@ -10,16 +10,16 @@ import urllib3
 from typing import TYPE_CHECKING, Dict, List, Optional
 
 from skyplane import compute
-from skyplane.api.impl.tracker import TransferProgressTracker
-from skyplane.api.impl.transfer_job import CopyJob, SyncJob, TransferJob
-from skyplane.api.transfer_config import TransferConfig
+from skyplane.api.tracker import TransferProgressTracker
+from skyplane.api.transfer_job import CopyJob, SyncJob, TransferJob
+from skyplane.api.config import TransferConfig
 from skyplane.replicate.replication_plan import ReplicationTopology, ReplicationTopologyGateway
 from skyplane.utils import logger
 from skyplane.utils.definitions import gateway_docker_image
 from skyplane.utils.fn import PathLike, do_parallel
 
 if TYPE_CHECKING:
-    from skyplane.api.impl.provisioner import Provisioner
+    from skyplane.api.provision.provisioner import Provisioner
 
 
 class DataplaneAutoDeprovision:
